requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
Django==5.1.3
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
python-telegram-bot==21.7
sniffio==1.3.1
sqlparse==0.5.2
